Colour calendar events based on whether release is closed, open, or overdue
JIRA and GitHub both provide status information in their API responses
so it is possible to unify that into a single model and render events
differently according to their status.

This change adds the status field as an enum in Release, and then
sets the background colour of the rendered event, green for released
and red for open but overdue.

See gh-6