Add status field to Release and colourize the calender events
JIRA and Github both provide status information in their API responses
so it is possible to unify that into a single model and render events
differently accroding to their status.

This change adds the status field as an enum in Release, and then
sets the background colour of the rendered event, green for released
and red for open but past due.