Treat releases from iCal (Google Calendar) has having unknown status
Google Calendar doesn't provide a way to mark an event on a calendar
has being completed or closed. This meant that all releases from a
Google Calendar had a status of OPEN and that they would be flagged
as overdue once the release date had passed, irrespective of whether
or not the release had actually happened.

This commit updates the handling of releases from iCal to always
treat their status as unknown.

Closes gh-11